requirements.txt
scrapy
scrapy-user-agents
pyyaml
pandas